Resolved issue misreading MH/OH weapons
The routine I was using returned the slot # instead of the Name of the item in the 'Main' and 'Sub' gear slots. It masked my mistake by changing back gear when expected, but it really wasn't comparing gear name. That is fixed now. Also dealt with a timing issue. When you zone, for awhile the gear is detected as missing. This is just a transition state that goes away once the zoning is complete, but it generated errors. I have mitigated the problem entries.